Avoid sending full `Book.page`; reduce network use
Before, when calling eg. `Book.make_and_save` on the client side, we
would provide a parameter `contents_and_parameters` carrying a list of
`Book.page = BookCore.page`. This type include full `VersionCore.t` and
`SetCore.t`, making for a potentially big request on the network.

This commit changes this to ensure that what travels on the network is
what was called before `BookCore.page_slug` and is now called
`BookCore.PageCore.t` for consistency (and verbosity \o). This makes for
lighter requests, but not such satisfactory code...