Get rid of `for-dance` in version and set parameters
In books in the database, rewrite them as dance pages, with a version or
with a set. There is a bunch of follow-up work to be done to show those
pages nicely in the book PDF, as for now dances are just happily
ignored. Similarly, the interface is still a bit rough around the edges,
but I am not sure what will be needed, and this therefore requires some
time playing with it to know what and how to improve.