Full rework of the renderer (#600)
This PR reworks the controllers entirely. It introduces a different
module, the “renderer”. The renderer is a mixed piece of code, with a
Nix part handling the compilation from a few values to full blown
`.svg`, `.ogg` and `.pdf` files, a piece of OCaml providing the
interface to Nix, and a piece of OCaml to go from models and parameters
to the JSON interface.

The result is, sadly, that the compilation is now much slower. However,
the resulting files are much cleaner, the compilation is now safe
(curtesy of the Nix sandbox), and:

- Partially addresses #23, #93 and #94 by removing a bunch of caches.
- Closes #31 by deferring this work to Nix
- Closes #78 by removing the notion of forced set pages
- Closes #81 by providing most of the currently-supported options in the
dialogs
- Closes #82 by having TeX take care of the compression in a simpler way
- Closes #198 by deferring this work to Nix
- Closes #447 by cleaning up the code
- Closes #470 by adding the option “show headers and footers” to set and
book download dialogs
- Closes #471 by adding the option “show headers and footers” to set and
book download dialogs
- Closes #472 by providing most of the currently-supported options in
the dialogs
- Closes #574 by having the new version support all that is necessary

There is a chance that this introduces breakages, but we will fix that
later.