Discontinue `fancyUnfoldRepeats`
The maintenance burden is too big and it has been causing weird errors
for many users. Of course, without it, a lot of playbacks will look
weird, because we lazy people tend not to duplicate the repeats in the
chords. This will not be such a problem for destructured versions, and
the rest can just be fixed by improving the database quality.