Merge `BuildSvg` and `BuildOgg` into one common `BuildSnippets` endpoint
The resulting code in the client is getting uglier and uglier, but it
only needs the right helpers. As of now, we do not make great use of
this, because we could avoid polling the job API twice all the time,
since it is the same actual job behind. One clean solution would be to
really support pairs of job + output file, but an easier solution would
be to craft an OGG promise that only starts polling when the SVG one is
done. Still, this is a step in the right direction, and we can merge it
as-is.